% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{calculate_cadence}
\alias{calculate_cadence}
\title{Calculate cadence from initial contacts}
\usage{
calculate_cadence(data, initial_contacts, sampling_rate_hz = 100, ...)
}
\arguments{
\item{data}{Sensor samples as an R data frame or Python pandas data frame.}

\item{initial_contacts}{Initial-contact data frame.}

\item{sampling_rate_hz}{Sampling frequency in Hertz.}

\item{...}{Arguments passed to the Python cadence constructor.}
}
\value{
A data frame of per-second cadence estimates.
}
\description{
Calculate cadence from initial contacts
}
\examples{
\dontrun{calculate_cadence(wrist_imu, initial_contacts)}
}
