% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{have_multigait}
\alias{have_multigait}
\title{Test whether MultiGait is available}
\usage{
have_multigait()
}
\value{
A single logical value. \code{FALSE} is returned when Python or the
\code{multigait} Python package cannot be initialized.
}
\description{
Test whether MultiGait is available
}
\examples{
have_multigait()
}
