% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{initial_contact_detector}
\alias{initial_contact_detector}
\title{Create an initial-contact detector}
\usage{
initial_contact_detector(
  algorithm = c("zijlstra", "mccamley", "pham", "ducharme", "gu", "mico_amigo", "sassi"),
  ...
)
}
\arguments{
\item{algorithm}{One of \code{"zijlstra"}, \code{"mccamley"}, \code{"pham"},
\code{"ducharme"}, \code{"gu"}, \code{"mico_amigo"}, or \code{"sassi"}.}

\item{...}{Arguments passed to the selected Python detector.}
}
\value{
A Python MultiGait initial-contact detector.
}
\description{
Create an initial-contact detector
}
