% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{interpolate_multigait}
\alias{interpolate_multigait}
\title{Interpolate sensor time series}
\usage{
interpolate_multigait(data, sample_rate = NULL, overlap_windows = FALSE)
}
\arguments{
\item{data}{A data frame, or list of data frames, to interpolate.}

\item{sample_rate}{Sampling frequency in Hertz. When \code{NULL}, it is obtained
from \code{data} with \code{\link[actibase:get_sample_rate]{actibase::get_sample_rate()}}.}

\item{overlap_windows}{Whether source windows overlap.}
}
\value{
A list of interpolated data frames.
}
\description{
Interpolate sensor time series
}
