% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{multigait_module}
\alias{multigait_module}
\title{Import the MultiGait Python package}
\usage{
multigait_module(convert = FALSE)
}
\arguments{
\item{convert}{Should reticulate convert Python objects to R objects?}
}
\value{
The imported Python module.
}
\description{
Import the MultiGait Python package
}
