% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_multimorbidity_impaired_pipeline}
\alias{run_multimorbidity_impaired_pipeline}
\title{Run the multimorbidity-impaired pipeline}
\usage{
run_multimorbidity_impaired_pipeline(data, safe = TRUE)
}
\arguments{
\item{data}{A dataset created by \code{\link[=as_multigait_data]{as_multigait_data()}} or a compatible
Python dataset object.}

\item{safe}{If \code{TRUE}, use the upstream \code{safe_run()} method.}
}
\value{
The result list returned by \code{\link[=run_multigait]{run_multigait()}}.
}
\description{
Run the multimorbidity-impaired pipeline
}
