% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{detect_gait_sequences}
\alias{detect_gait_sequences}
\title{Detect gait sequences}
\usage{
detect_gait_sequences(data, detector = gait_sequence_detector())
}
\arguments{
\item{data}{Sensor samples as an R data frame or Python pandas data frame.}

\item{detector}{A detector returned by \code{\link[=gait_sequence_detector]{gait_sequence_detector()}}.}
}
\value{
A data frame of detected gait sequences.
}
\description{
Detect gait sequences
}
\examples{
\dontrun{detect_gait_sequences(wrist_imu)}
}
