% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{detect_gait_sequences}
\alias{detect_gait_sequences}
\title{Detect gait sequences}
\usage{
detect_gait_sequences(data, detector = gait_sequence_detector())
}
\arguments{
\item{data}{Sensor samples as an R data frame or Python pandas data frame.}

\item{detector}{A detector returned by \code{\link[=gait_sequence_detector]{gait_sequence_detector()}}.}
}
\value{
A data frame of detected gait sequences.
}
\description{
Detect gait sequences
}
\examples{
component_imu <- wrist_imu[c("acc_x", "acc_y", "acc_z")]
names(component_imu) <- c("acc_is", "acc_ml", "acc_pa")
gait_sequences <- detect_gait_sequences(
  component_imu, gait_sequence_detector("ionescu", version = "wrist")
)
head(gait_sequences)
}
