% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{detect_initial_contacts}
\alias{detect_initial_contacts}
\title{Detect initial contacts}
\usage{
detect_initial_contacts(data, detector = initial_contact_detector())
}
\arguments{
\item{data}{Sensor samples as an R data frame or Python pandas data frame.}

\item{detector}{An initial-contact detector from \code{\link[=initial_contact_detector]{initial_contact_detector()}}.}
}
\value{
A data frame of initial contacts.
}
\description{
Detect initial contacts
}
\examples{
component_imu <- wrist_imu[c("acc_x", "acc_y", "acc_z")]
names(component_imu) <- c("acc_is", "acc_ml", "acc_pa")
initial_contacts <- detect_initial_contacts(
  component_imu, initial_contact_detector("zijlstra", version = "wrist")
)
head(initial_contacts)
}
