% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{interpolate_multigait}
\alias{interpolate_multigait}
\title{Interpolate sensor time series}
\usage{
interpolate_multigait(data, sampling_rate_hz = 100, overlap_windows = FALSE)
}
\arguments{
\item{data}{A data frame, or list of data frames, to interpolate.}

\item{sampling_rate_hz}{Sampling frequency in Hertz.}

\item{overlap_windows}{Whether source windows overlap.}
}
\value{
A list of interpolated data frames.
}
\description{
Interpolate sensor time series
}
